move methods to support more readers
