+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/segment/index/readers/forward/FixedByteChunkSVForwardIndexReader.java
@@ -19,10 +19,6 @@
 package org.apache.pinot.segment.local.segment.index.readers.forward;
 
 import java.nio.ByteBuffer;
-import java.nio.DoubleBuffer;
-import java.nio.FloatBuffer;
-import java.nio.IntBuffer;
-import java.nio.LongBuffer;
 import javax.annotation.Nullable;
 import org.apache.pinot.segment.local.io.writer.impl.FixedByteChunkSVForwardIndexWriter;
 import org.apache.pinot.segment.spi.memory.PinotDataBuffer;
@@ -49,175 +45,6 @@ public final class FixedByteChunkSVForwardIndexReader extends BaseChunkSVForward
       return new ChunkReaderContext(_chunkSize);
     } else {
       return null;
-    }
-  }
-
-  @Override
-  public void readValuesSV(int[] docIds, int length, int[] values, ChunkReaderContext context) {
-    if (getValueType().getStoredType().isFixedWidth() && !_isCompressed && isContiguousRange(docIds, length)) {
-      switch (getValueType().getStoredType()) {
-        case INT: {
-          int minOffset = docIds[0] * Integer.BYTES;
-          IntBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Integer.BYTES).asIntBuffer();
-          buffer.get(values, 0, length);
-        }
-        break;
-        case LONG: {
-          int minOffset = docIds[0] * Long.BYTES;
-          LongBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Long.BYTES).asLongBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = (int) buffer.get(i);
-          }
-        }
-        break;
-        case FLOAT: {
-          int minOffset = docIds[0] * Float.BYTES;
-          FloatBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Float.BYTES).asFloatBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = (int) buffer.get(i);
-          }
-        }
-        break;
-        case DOUBLE: {
-          int minOffset = docIds[0] * Double.BYTES;
-          DoubleBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Double.BYTES).asDoubleBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = (int) buffer.get(i);
-          }
-        }
-        break;
-        default:
-          throw new IllegalArgumentException();
-      }
-    } else {
-      super.readValuesSV(docIds, length, values, context);
-    }
-  }
-
-  @Override
-  public void readValuesSV(int[] docIds, int length, long[] values, ChunkReaderContext context) {
-    if (getValueType().getStoredType().isFixedWidth() && !_isCompressed && isContiguousRange(docIds, length)) {
-      switch (getValueType().getStoredType()) {
-        case INT: {
-          int minOffset = docIds[0] * Integer.BYTES;
-          IntBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Integer.BYTES).asIntBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = buffer.get(i);
-          }
-        }
-        break;
-        case LONG: {
-          int minOffset = docIds[0] * Long.BYTES;
-          LongBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Long.BYTES).asLongBuffer();
-          buffer.get(values, 0, length);
-        }
-        break;
-        case FLOAT: {
-          int minOffset = docIds[0] * Float.BYTES;
-          FloatBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Float.BYTES).asFloatBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = (long) buffer.get(i);
-          }
-        }
-        break;
-        case DOUBLE: {
-          int minOffset = docIds[0] * Double.BYTES;
-          DoubleBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Double.BYTES).asDoubleBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = (long) buffer.get(i);
-          }
-        }
-        break;
-        default:
-          throw new IllegalArgumentException();
-      }
-    } else {
-      super.readValuesSV(docIds, length, values, context);
-    }
-  }
-
-  @Override
-  public void readValuesSV(int[] docIds, int length, float[] values, ChunkReaderContext context) {
-    if (getValueType().getStoredType().isFixedWidth() && !_isCompressed && isContiguousRange(docIds, length)) {
-      switch (getValueType().getStoredType()) {
-        case INT: {
-          int minOffset = docIds[0] * Integer.BYTES;
-          IntBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Integer.BYTES).asIntBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = buffer.get(i);
-          }
-        }
-        break;
-        case LONG: {
-          int minOffset = docIds[0] * Long.BYTES;
-          LongBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Long.BYTES).asLongBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = buffer.get(i);
-          }
-        }
-        break;
-        case FLOAT: {
-          int minOffset = docIds[0] * Float.BYTES;
-          FloatBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Float.BYTES).asFloatBuffer();
-          buffer.get(values, 0, length);
-        }
-        break;
-        case DOUBLE: {
-          int minOffset = docIds[0] * Double.BYTES;
-          DoubleBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Double.BYTES).asDoubleBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = (float) buffer.get(i);
-          }
-        }
-        break;
-        default:
-          throw new IllegalArgumentException();
-      }
-    } else {
-      super.readValuesSV(docIds, length, values, context);
-    }
-  }
-
-  @Override
-  public void readValuesSV(int[] docIds, int length, double[] values, ChunkReaderContext context) {
-    if (getValueType().getStoredType().isFixedWidth() && !_isCompressed && isContiguousRange(docIds, length)) {
-      switch (getValueType().getStoredType()) {
-        case INT: {
-          int minOffset = docIds[0] * Integer.BYTES;
-          IntBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Integer.BYTES).asIntBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = buffer.get(i);
-          }
-        }
-        break;
-        case LONG: {
-          int minOffset = docIds[0] * Long.BYTES;
-          getLong(0, context);
-          LongBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Long.BYTES).asLongBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = buffer.get(i);
-          }
-        }
-        break;
-        case FLOAT: {
-          int minOffset = docIds[0] * Float.BYTES;
-          FloatBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Float.BYTES).asFloatBuffer();
-          for (int i = 0; i < buffer.limit(); i++) {
-            values[i] = buffer.get(i);
-          }
-        }
-        break;
-        case DOUBLE: {
-          int minOffset = docIds[0] * Double.BYTES;
-          DoubleBuffer buffer = _rawData.toDirectByteBuffer(minOffset, length * Double.BYTES).asDoubleBuffer();
-          buffer.get(values, 0, length);
-        }
-        break;
-        default:
-          throw new IllegalArgumentException();
-      }
-    } else {
-      super.readValuesSV(docIds, length, values, context);
     }
   }
 
@@ -263,9 +90,5 @@ public final class FixedByteChunkSVForwardIndexReader extends BaseChunkSVForward
     } else {
       return _rawData.getDouble(docId * Double.BYTES);
     }
-  }
-
-  private boolean isContiguousRange(int[] docIds, int length) {
-    return docIds[length - 1] - docIds[0] == length - 1;
   }
 }
